Improve `pd.Series` initialization

+++ tests/unit_tests/analysis/test_analysis_statistics_profit_factor.py
@@ -14,6 +14,7 @@
 # -------------------------------------------------------------------------------------------------
 
 import pandas as pd
+from numpy import float64
 
 from nautilus_trader.analysis.statistics.profit_factor import ProfitFactor
 
@@ -32,7 +33,7 @@ class TestProfitFactorPortfolioStatistic:
     def test_calculate_given_empty_series_returns_nan(self):
         # Arrange
         stat = ProfitFactor()
-        data = pd.Series([0.0])
+        data = pd.Series([0.0], dtype=float64)
 
         # Act
         result = stat.calculate_from_returns(data)
@@ -43,7 +44,7 @@ class TestProfitFactorPortfolioStatistic:
     def test_calculate_given_mix_of_pnls_returns_expected(self):
         # Arrange
         stat = ProfitFactor()
-        data = pd.Series([3.0, 2.0, 1.0, -1.0, -2.0])
+        data = pd.Series([3.0, 2.0, 1.0, -1.0, -2.0], dtype=float64)
 
         # Act
         result = stat.calculate_from_returns(data)
